Nesting Configuration within IndexStore space
